refactored frameworks/[id] and StepView to use a common component
